Removes mtl parsing, adds 'grayscale' shader, adds key inputs

+++ src/utils/vector.c
@@ -6,7 +6,7 @@
 /*   By: chaueur <chaueur@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/11 18:06:38 by chaueur           #+#    #+#             */
-/*   Updated: 2016/05/24 12:29:40 by chaueur          ###   ########.fr       */
+/*   Updated: 2016/05/26 11:05:16 by chaueur          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,14 +47,14 @@ t_vec				*vec_cross_prod(t_vec *v1, t_vec *v2)
 
 void				vec_normalize(t_vec *v)
 {
-	float		v_len;
+	double			v_len;
 
 	v_len = sqrt(v->x * v->x + v->y * v->y + v->z * v->z);
 	if (v_len)
 	{
-		v->x /= v_len;
-		v->y /= v_len;
-		v->z /= v_len;
+		v->x /= (float)v_len;
+		v->y /= (float)v_len;
+		v->z /= (float)v_len;
 	}
 }
